Integrated Assimp, constructed basic render manager framework

+++ amGraphicsDX/Source/amDXInputLayout.cpp
@@ -8,7 +8,7 @@ namespace amEngineSDK {
 
   amDXInputLayout::~amDXInputLayout() {}
   int32 amDXInputLayout::Create(amDXDevice* _pDV, ID3DBlob* _pShaderBlob) {
-    _pDV->m_pDV->CreateInputLayout(&m_layout[0], m_layout.size(), _pShaderBlob->GetBufferPointer(),
+    _pDV->m_pDV->CreateInputLayout(&m_layout[0], static_cast<uint32>(m_layout.size()), _pShaderBlob->GetBufferPointer(),
                                    _pShaderBlob->GetBufferSize(), &m_pVertexLayout);
     return 0;
   }
